fix network issues
